feat: add initial metadata evaluation
